% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_test.R
\name{combination.test}
\alias{combination.test}
\title{Compute the global p-value using the heavy-tailed combination test}
\usage{
combination.test(
  p.vec,
  weights = NA,
  method = "Cauchy",
  tail.idx = 1,
  truncate = F,
  truncate.threshold = 0.99
)
}
\arguments{
\item{p.vec}{an vector of base p-values. Missing values are allowed.}

\item{weights}{an vector of non-negative weights for each base hypothesis.
When it's missing, all hypotheses are assigned equal weights.}

\item{method}{one of "Cauchy", "Log Cauchy", "Levy", "Paret", "Frechet", "t", or "Inverse Gamma", with the default "Cauchy".}

\item{tail.idx}{is the shape parameter for Cauchy, Pareto, Frechet, t, and Inverse Gamma.}

\item{truncate}{logical value, default is FALSE.
If \code{truncate} is TRUE, then apply truncated Cauchy instead of standard Cauchy.}

\item{truncate.threshold}{truncated threshold (upper bound) for the p-values when \code{truncate} is TRUE.}
}
\value{
a p-value for the global test
}
\description{
Compute the global p-value using the heavy-tailed combination test
}
