% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{data_Gen}
\alias{data_Gen}
\title{Generate p-value matrix for simulation}
\usage{
data_Gen(n, p = 2, mu, rho = 0, copula = "gaussian", dof = 5, one_sided = F)
}
\arguments{
\item{n}{number of repeated times}

\item{p}{number of base hypotheses}

\item{mu}{a vector of possible values of effect size of nonnull individual
hypotheses}

\item{rho}{a common correlation between base hypotheses}

\item{copula}{base test statistics from multivariate gaussian or multivariate t}

\item{dof}{the degree of freedom of the multivariate t}

\item{one.sided}{either construct one-sided p-value or not}
}
\value{
a list of objects
\describe{
\item{p.mat}{the p-value matrix of size \code{n * p} where each column
corresponding to one base hypothesis}
\item{z.mat}{The corresponding base test statistics matrix of size \code{n * p},
which are either z-scores or t-scores}
}
}
\description{
Generate p-value matrix for simulation
}
