% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_test.R
\docType{package}
\name{heavytailcombtest_package}
\alias{heavytailcombtest_package}
\title{Compute the global p-value using the heavy-tailed combination test.}
\description{
Computes a global p-value using the heavy-tailed combination test.
The function \code{combination.test} realizes all three versions---standard, average, and weighted versions---of the combination test.
Specific choices also correspond to the Cauchy combination test and harmonic mean p-value method.
}
