% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pfrechet}
\alias{pfrechet}
\title{The Frechet Distribution}
\usage{
pfrechet(x, m=0, s=1, alpha=1, lower.tail = T)

qfrechet(p, m=0, s=1, alpha=1)
}
\arguments{
\item{x}{vector of quantiles}

\item{m}{vectors of location parameters}

\item{s}{vectors of scale parameters}

\item{alpha}{vectors of shape parameters}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X\leq x); otherwise, P(X>x)}

\item{p}{vectors of probabilities}
}
\description{
Distribution and quantile function for the Frechet distribution
with the location parameter equal to \code{m}, scale parameter equal to \code{s}, and shape parameter equal to \code{alpha}.
}
