% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ppareto}
\alias{ppareto}
\title{The Pareto Distribution}
\usage{
ppareto(x, x_m = 1, alpha = 1, lower.tail = T)

qpareto(p, x_m = 1, alpha = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{x_m}{vectors of location parameters}

\item{alpha}{vectors of shape parameters}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X\le x)}; otherwise, \eqn{P(X>x)}.}

\item{p}{vectors of probabilities}
}
\description{
Distribution and quantile function for the Pareto distribution
with the location parameter equal to \code{x_m} and shape parameter equal to \code{alpha}.
}
