% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ptcauchy}
\alias{ptcauchy}
\title{The truncated Cauchy Distribution}
\usage{
plcauchy(x, lower.tail = T)

qlcauchy(p)
}
\arguments{
\item{x}{vector of quantiles}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X\le x)}; otherwise, \eqn{P(X>x)}}

\item{p}{vectors of probabilities}
}
\description{
Distribution and quantile function for the truncated Cauchy distribution
with the trunacted threshold (lower bound) \code{threshold}
}
