% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qlcauchy}
\alias{qlcauchy}
\title{The Log Cauchy Distribution}
\usage{
plcauchy(x, lower.tail = T)

qlcauchy(p)
}
\arguments{
\item{p}{vectors of probabilities}

\item{x}{vector of quantiles}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X\le x)}; otherwise, \eqn{P(X>x)}.}
}
\description{
Distribution and quantile function for the log Cauchy distribution
}
